pyproject.toml
[project]
name = "bouncer"
version = "0.0.0"
description = "A Discord moderation bot to keep track of the bad eggs "
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "humanize>=4.10.0",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "pyyaml>=6.0.2",
]

[tool.pyright]
# reportAny = false
reportUnusedCallResult = false
